+++ Diagnostics/Assets/Scripts/KLib/Signals/Level.cs
@@ -147,7 +147,6 @@ namespace KLib.Signals
 
         public void ResetSweepables()
         {
-            Debug.Log($"level = {Value}");
             _lastAtten = float.NaN;
             if (Units == LevelUnits.mA)
             {